Launch the language server via a `WasmProcess`
Copying from [1] and [2] this commit implements the
`createServerProcess` method to configure a `WasmProcess` instance to
launch the WASI build of Python to run the language server.

The mount point `kind: 'workspaceFolder'` makes the user's workspace
available to the server process under the `/workspace` directory.

The mount point `kind: 'extensionLocation' makes a subdirectory of the
extension's source code available to the server process under a
directory of your choosing. Here we use it to make both Python's
standard library and any 3rd party packages we installed available to
the process, in the location it expects to see.

Lastly the `process.stderr.onData` callback can be used to forward the
process' `stderr` stream to a VSCode output window. Unfortunately,
however I could not make this work on `vscode.dev` - something to do
with the Cross-Origin Isolation not liking the underlying
SharedBufferArray...

[1]: https://github.com/microsoft/vscode-wasm/blob/65669200000306d174ce2bbfdd9e4d41e9517466/testbeds/lsp-rust/client/src/extension.ts#L18-L48
[2]: https://github.com/microsoft/vscode-wasm/blob/65669200000306d174ce2bbfdd9e4d41e9517466/testbeds/python/extension.ts#L17-L38