Implement the web version of the extension
Copying from [1], this connects up the generic client we wrote to the
VSCode's `LanguageClient` implementation

Most of the plumbing is almost identical to the node version, with
most differences being just the import paths.

The only slight difference to the node version was having to write a
custom subclass of `BaseLanguageClient` that takes a `WasmProcess` as
opposed to a `Worker`

This also implements the web-specific VSCode extension entry point.

[1]: https://github.com/microsoft/vscode-wasm/blob/65669200000306d174ce2bbfdd9e4d41e9517466/testbeds/lsp-rust/client/src/lspServer.ts